Move the online tests to memory `tests.rs`
